refactor internal scalebar composables to ScalebarRenderer.kt
